+++ src/lib.rs
@@ -59,6 +59,17 @@ impl Load for Inventory {
 }
 
 impl Inventory {
+    pub fn get_host(self, hostname: &str) -> Result<Host> {
+        let hostvars: &Value = &self.data["_meta"]["hostvars"][hostname];
+        let host = Host {
+            name: hostname.to_string(),
+            data: hostvars.to_owned(),
+        };
+        Ok(host)
+    }
+}
+
+impl Inventory {
     pub fn get_hosts(self) -> Result<Vec<Host>> {
         let mut hosts = Vec::new();
         for (hostname, hostvars) in self.data["_meta"]["hostvars"].as_object().unwrap() {